+++ packages/manager/apps/web/client/app/hosting/email/activate/activate.component.js
@@ -1,0 +1,13 @@
+import controller from './activate.controller';
+import template from './activate.html';
+
+export default {
+  bindings: {
+    serviceName: '<',
+    user: '<',
+    domainNames: '<',
+    goToHosting: '<',
+  },
+  controller,
+  template,
+};